refactor TsetControls, ComingSoon, new year date into separate files
