+++ src/lib/constants.js
@@ -1,0 +1,2 @@
+export const NEW_YEAR_DATE = "2025-01-01T00:00:00";
+// export const NEW_YEAR_DATE = "2023-12-31T18:28:00";
